refactor folders and gaps into files
