+++ src/components/modal/modal.jsx
@@ -3,7 +3,7 @@ import { createPortal } from 'react-dom'
 import { useEffect } from 'react'
 
 import { CloseIcon } from '@ya.praktikum/react-developer-burger-ui-components'
-import ModalOverlay from './modal-overlay/modal-overlay'
+import ModalOverlay from '../modal-overlay/modal-overlay'
 
 import styles from './modal.module.css'
